requirements.txt
pandas
numpy
matplotlib
seaborn
pyspark==3.2.4
scikit-learn==1.7.2
streamlit
fastapi
uvicorn
psycopg2-binary
requests
